fix: corregir bugs en models y agregar MMR-DETR wrapper
- ssd_mobilenet: eliminar normalización incorrecta de cajas en forward(),
  reemplazar _device fijo por property dinámica
- mmr_detr_wrapper + rt_detr_wrapper: retornos de batch vacío ahora usan
  requires_grad=True y device correcto para evitar errores en backward()
- yolo26_wrapper: dtype=torch.float32 explícito en tensores vacíos
- registry: separar TypeError/AttributeError, warning en lugar de fallo silencioso
- Agregar MMR-DETR wrapper, configs y actualizar CLI/trainer

+++ model/models/yolo26_wrapper.py
@@ -665,9 +665,9 @@ class YOLO26Detector(BaseDetector):
 
             if boxes is None or len(boxes) == 0:
                 predictions.append({
-                    "boxes": torch.zeros((0, 4), device=device),
+                    "boxes": torch.zeros((0, 4), dtype=torch.float32, device=device),
                     "labels": torch.zeros((0,), dtype=torch.int64, device=device),
-                    "scores": torch.zeros((0,), device=device),
+                    "scores": torch.zeros((0,), dtype=torch.float32, device=device),
                 })
                 continue
 
@@ -683,9 +683,9 @@ class YOLO26Detector(BaseDetector):
 
             if xyxy.shape[0] == 0:
                 predictions.append({
-                    "boxes": torch.zeros((0, 4), device=device),
+                    "boxes": torch.zeros((0, 4), dtype=torch.float32, device=device),
                     "labels": torch.zeros((0,), dtype=torch.int64, device=device),
-                    "scores": torch.zeros((0,), device=device),
+                    "scores": torch.zeros((0,), dtype=torch.float32, device=device),
                 })
             else:
                 predictions.append({
